+++ react/src/components/search.tsx
@@ -7,12 +7,11 @@ import { searchIconLink, typesInOrder } from '../navigation/links'
 import { useColors } from '../page_template/colors'
 import { useSettings } from '../page_template/settings'
 import '../common.css'
-import { SearchResult, SearchParams, getIndexCacheKey, SearchIndexConfig } from '../search'
-import { SearchWorkerInputMessage, SearchWorkerOutputMessage, SearchWorkerStatus } from '../searchWorker'
+import { SearchResult, SearchParams, debugPerformance, getIndexCacheKey, SearchIndexConfig } from '../search'
+import type { SearchWorkerInputMessage, SearchWorkerOutputMessage, SearchWorkerStatus } from '../searchWorker'
 import { Universe, useUniverse } from '../universe'
 import { Property } from '../utils/Property'
 import { TestUtils } from '../utils/TestUtils'
-import { debugPerformance } from '../utils/search-performance'
 
 import { GenericSearchBox } from './search-generic'
