--- react/src/utils/search-performance.ts
@@ -1,10 +1,0 @@
-const debugSearchPerformance: boolean = false
-
-export function debugPerformance(arg: unknown): void {
-    if (debugSearchPerformance) {
-        // eslint-disable-next-line no-console -- Debug logging
-        console.log(arg)
-    }
-}
-
-export const simulateSlowSearchDelayMs: number | undefined = undefined
